Refactor image adjustments. Easier to create plugins
Start implementing config options for video encoding
