+++ plugins/convert/_config.py
@@ -4,6 +4,7 @@
 import logging
 
 from lib.config import FaceswapConfig
+from lib.utils import _video_extensions
 
 logger = logging.getLogger(__name__)  # pylint: disable=invalid-name
 
@@ -27,7 +28,7 @@ class Config(FaceswapConfig):
 #                         info="Options that apply to all models")
 
         # << BOX OPTIONS >> #
-        section = "box.blend"
+        section = "mask.box_blend"
         self.add_section(title=section,
                          info="Options for blending the edges of the swapped box with the "
                               "background image")
@@ -63,7 +64,7 @@ class Config(FaceswapConfig):
                  "this too high")
 
         # << MASK OPTIONS >> #
-        section = "mask.blend"
+        section = "mask.mask_blend"
         self.add_section(title=section,
                          info="Options for blending the edges between the mask and the "
                               "background image")
@@ -86,9 +87,15 @@ class Config(FaceswapConfig):
                  "type blending."
                  "\nAdditional passes have exponentially less effect so it's not worth setting "
                  "this too high")
+        self.add_item(
+            section=section, title="erosion", datatype=float, default=0.0, rounding=1,
+            min_max=(-100.0, 100.0),
+            info="Erosion kernel size as a percentage of the mask radius area.\n"
+                 "Positive values apply erosion which reduces the size of the swapped area.\n"
+                 "Negative values apply dilation which increases the swapped area")
 
         # << COLOR OPTIONS >> #
-        section = "face.color_transfer"
+        section = "color.color_transfer"
         self.add_section(title=section,
                          info="Options for transfering the color distribution from the source to "
                               "the target image using the mean and standard deviations of the "
@@ -112,7 +119,7 @@ class Config(FaceswapConfig):
                  "scaling factor proposed in the paper. This method seems to produce more "
                  "consistently aesthetically pleasing results")
 
-        section = "face.match_hist"
+        section = "color.match_hist"
         self.add_section(title=section,
                          info="Options for matching the histograms between the source and "
                               "destination faces")
@@ -123,12 +130,12 @@ class Config(FaceswapConfig):
                  "in by filtering out colors at the extreme ends of the histogram spectrum")
 
         # << POST WARP OPTIONS >> #
-        section = "scaling.sharpen_image"
+        section = "scaling.sharpen"
         self.add_section(title=section,
                          info="Options for sharpening the face after placement")
         self.add_item(
             section=section, title="method", datatype=str,
-            choices=["none", "box", "gaussian", "unsharp_mask"], default="none",
+            choices=["box", "gaussian", "unsharp_mask"], default="unsharp_mask",
             info="The type of sharpening to use: "
                  "\n\t box: Fastest, but weakest method. Uses a box filter to assess edges."
                  "\n\t gaussian: Slower, but better than box. Uses a gaussian filter to assess "
@@ -164,3 +171,94 @@ class Config(FaceswapConfig):
                  "while leaving subtler edges untouched. "
                  "\nLow values should sharpen more because fewer areas are excluded. "
                  "\nHigher threshold values exclude areas of lower contrast.")
+
+        # << VIDEO ENCODING OPTIONS >> #
+        section = "video.global"
+        self.add_section(title=section,
+                         info="Options for encoding converted videos")
+        self.add_item(
+            section=section, title="container", datatype=str, default="mp4",
+            choices=[ext.replace(".", "") for ext in _video_extensions],
+            info="Video container to use.")
+        self.add_item(
+            section=section, title="codec", datatype=str,
+            choices=["libx264", "libx265"], default="libx264",
+            info="Video codec to use:\n"
+                 "libx264: H.264. A widely supported and commonly used codec.\n"
+                 "libx265: H.265 / HEVC video encoder application library.")
+
+        section = "video.libx264"
+        self.add_section(title=section,
+                         info="Options for h.264 encoding")
+        self.add_item(
+            section=section, title="crf", datatype=int, min_max=(0, 51), rounding=1, default=23,
+            info="Constant Rate Factor:  0 is lossless and 51 is worst quality possible. A lower "
+                 "value generally leads to higher quality, and a subjectively sane range is "
+                 "17–28. Consider 17 or 18 to be visually lossless or nearly so; it should look "
+                 "the same or nearly the same as the input but it isn't technically lossless.\n"
+                 "The range is exponential, so increasing the CRF value +6 results in roughly "
+                 "half the bitrate / file size, while -6 leads to roughly twice the bitrate.\n"
+                 "Choose the highest CRF value that still provides an acceptable quality. If the "
+                 "output looks good, then try a higher value. If it looks bad, choose a lower "
+                 "value.")
+        self.add_item(
+            section=section, title="preset", datatype=str, default="medium",
+            choices=["ultrafast", "superfast", "veryfast", "faster", "fast", "medium", "slow",
+                     "slower", "veryslow"],
+            info="A preset is a collection of options that will provide a certain encoding speed "
+                 "to compression ratio. A slower preset will provide better compression "
+                 "(compression is quality per filesize). Use the slowest preset that you have "
+                 "patience for")
+        self.add_item(
+            section=section, title="tune", datatype=str, default="none",
+            choices=["none", "film", "animation", "grain", "stillimage", "fastdecode",
+                     "zerolatency"],
+            info="Change settings based upon the specifics of your input:\n"
+                 "film: use for high quality movie content; lowers deblocking.\n\t"
+                 "animation: good for cartoons; uses higher deblocking and more reference "
+                 "frames.\n\t"
+                 "grain: preserves the grain structure in old, grainy film material.\n\t"
+                 "stillimage: good for slideshow-like content.\n\t"
+                 "fastdecode: allows faster decoding by disabling certain filters.\n\t"
+                 "zerolatency: good for fast encoding and low-latency streaming.")
+        self.add_item(
+            section=section, title="profile", datatype=str, default="auto",
+            choices=["auto", "baseline", "main", "high", "high10", "high422", "high444"],
+            info="Limit the output to a specific H.264 profile. Don't change this unless your "
+                 "target device only supports a certain profile.")
+        self.add_item(
+            section=section, title="level", datatype=str, default="auto",
+            choices=["auto", "1", "1b", "1.1", "1.2", "1.3", "2", "2.1", "2.2", "3", "3.1", "3.2",
+                     "4", "4.1", "4.2", "5", "5.1", "5.2", "6", "6.1", "6.2"],
+            info="Set the encoder level, Don't change this unless your target device only "
+                 "supports a certain level.")
+
+        section = "video.libx265"
+        self.add_section(title=section,
+                         info="Options for h.265 / HEVC encoding")
+        self.add_item(
+            section=section, title="crf", datatype=int, min_max=(0, 51), rounding=1, default=28,
+            info="Constant Rate Factor:  0 is lossless and 51 is worst quality possible. A lower "
+                 "value generally leads to higher quality, and a subjectively sane range is "
+                 "17–28. Consider 17 or 18 to be visually lossless or nearly so; it should look "
+                 "the same or nearly the same as the input but it isn't technically lossless.\n"
+                 "The range is exponential, so increasing the CRF value +6 results in roughly "
+                 "half the bitrate / file size, while -6 leads to roughly twice the bitrate.\n"
+                 "Choose the highest CRF value that still provides an acceptable quality. If the "
+                 "output looks good, then try a higher value. If it looks bad, choose a lower "
+                 "value.")
+        self.add_item(
+            section=section, title="preset", datatype=str, default="medium",
+            choices=["ultrafast", "superfast", "veryfast", "faster", "fast", "medium", "slow",
+                     "slower", "veryslow"],
+            info="A preset is a collection of options that will provide a certain encoding speed "
+                 "to compression ratio. A slower preset will provide better compression "
+                 "(compression is quality per filesize). Use the slowest preset that you have "
+                 "patience for")
+        self.add_item(
+            section=section, title="tune", datatype=str, default="none",
+            choices=["none", "grain", "fastdecode", "zerolatency"],
+            info="Change settings based upon the specifics of your input:\n"
+                 "grain: preserves the grain structure in old, grainy film material.\n\t"
+                 "fastdecode: allows faster decoding by disabling certain filters.\n\t"
+                 "zerolatency: good for fast encoding and low-latency streaming.")
